Reestructuramos el proyecto respetando el estandar de Django
