--- pedigroup/model/empanada.py
@@ -1,5 +1,0 @@
-class Empanada:
-
-    def __init__(self, type, restaurant):
-        self.type = type
-        self.restaurant = restaurant
